Changes on the description
- resolved the i2c problem (on the other hand it still has bugs)
- introduced a smart gestion of the elements on the screen (the display updates itself only when an element is created or modified)

+++ Attiny_slave/settings.h
@@ -2,7 +2,7 @@
 #include <Wire.h>
 
 int i, a, c, r;
-int ready_byte = 1;
+byte ready_byte = 1;
 enum types {
   text,
   command,
